refactor: [relax] remove todo components
